refactor(auth): Split questionary authorizers
